[async result] Cleanup branch for merge
